+++ data/deduplicator/test/factory.go
@@ -4,8 +4,8 @@ import (
 	"github.com/tidepool-org/platform/data"
 	"github.com/tidepool-org/platform/data/store"
 	"github.com/tidepool-org/platform/data/types/upload"
-	"github.com/tidepool-org/platform/id"
 	"github.com/tidepool-org/platform/log"
+	"github.com/tidepool-org/platform/test"
 )
 
 type CanDeduplicateDatasetOutput struct {
@@ -41,7 +41,7 @@ type NewRegisteredDeduplicatorForDatasetOutput struct {
 }
 
 type Factory struct {
-	ID                                             string
+	*test.Mock
 	CanDeduplicateDatasetInvocations               int
 	CanDeduplicateDatasetInputs                    []*upload.Upload
 	CanDeduplicateDatasetOutputs                   []CanDeduplicateDatasetOutput
@@ -58,7 +58,7 @@ type Factory struct {
 
 func NewFactory() *Factory {
 	return &Factory{
-		ID: id.New(),
+		Mock: test.NewMock(),
 	}
 }